refactor: Fix norminette errors and organize logical components
