+++ src/validation/validate_properties.c
@@ -6,49 +6,30 @@
 /*   By: thaperei <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/04/05 13:29:00 by thaperei          #+#    #+#             */
-/*   Updated: 2026/04/13 18:45:26 by thaperei         ###   ########.fr       */
+/*   Updated: 2026/04/15 20:47:23 by thaperei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "validation.h"
 #include "libft.h"
 
-int	is_valid_ratio(char *str)
+int	is_valid_extension(char *file)
 {
-	float	value;
+	int		len;
+	char	*extension;
 
-	if (str == NULL || str[0] == '\0')
-		return (0);
-	value = ft_atof(str);
-	if (!is_valid_float(str) || value < 0.0 || value > 1.0)
-		return (0);
-	return (1);
-}
-
-int	is_valid_color(char *str)
-{
-	int		i;
-	int		value;
-	char	**colors;
-
-	if (str == NULL || str[0] == '\0')
-		return (0);
-	colors = ft_split(str, ',');
-	if (colors == NULL)
-		return (0);
-	i = -1;
-	while (colors[++i])
+	if (!file)
 	{
-		value = ft_atoi(colors[i]);
-		if (!is_valid_int(colors[i]) || value < 0 || value > 255)
-		{
-			free_arr(colors);
-			return (0);
-		}
-	}
-	free_arr(colors);
-	if (i != 3)
+		show_error("Invalid file");
 		return (0);
+	}
+	len = ft_strlen(file);
+	extension = ft_strrchr(file, '.');
+	if (!extension || len <= 3 || ft_strcmp(".rt", extension) != 0)
+	{
+		show_error("Invalid extension");
+		return (0);
+	}
 	return (1);
 }
